cleaning classes for low-level serialization
