+++ core/src/main/java/org/lionweb/lioncore/java/serialization/data/SerializedContainmentValue.java
@@ -3,6 +3,9 @@ package org.lionweb.lioncore.java.serialization.data;
 import java.util.List;
 import java.util.Objects;
 
+/**
+ * This represents the serialization of the values of a containment link in a Node.
+ */
 public class SerializedContainmentValue {
     private MetaPointer metaPointer;
     private List<String> value;
@@ -24,10 +27,16 @@ public class SerializedContainmentValue {
         this.metaPointer = metaPointer;
     }
 
+    /**
+     * This returns the list of Node-IDs contained.
+     */
     public List<String> getValue() {
         return value;
     }
 
+    /**
+     * This expects the list of Node-IDs contained.
+     */
     public void setValue(List<String> value) {
         this.value = value;
     }
